Models

EntropyScaling.ChapmanEnskogModel

EntropyScaling.FitOptions

EntropyScaling.FrameworkModel

EntropyScaling.RefpropRESModel

EntropyScaling.TransportPropertyData

EntropyScaling.Ω

Chapman-Enskog Models

Chapman-Enskog model the for transport properties at the zero-density limit based on the kinetic gas theory.

\[\begin{aligned} \eta_{\varrho \rightarrow 0} &= \frac{5}{16} \sqrt{\frac{M k_{\rm B} T}{\pi N_{\rm A}}} \frac{1}{\sigma^2 \Omega^{(2,2)}} \\ \lambda_{\varrho \rightarrow 0} &= \frac{75}{64} k_{\rm B} \sqrt{\frac{R T}{M \pi}} \frac{1}{\sigma^2 \Omega^{(2,2)}}\\ D_{\varrho \rightarrow 0} \varrho^{\rm m} &= \frac{3}{8} \sqrt{\frac{M k_{\rm B} T}{\pi N_{\rm A}}} \frac{1}{\sigma^2 \Omega^{(1,1)}} \end{aligned}\]

EntropyScaling.ChapmanEnskogModel — Type

ChapmanEnskogModel <: AbstractTransportPropertyModel

Chapman-Enskog transport properties for the zero-density limit.

Fields

σ::Vector{T}: Lennard-Jones size parameter ([σ] = m)

ε::Vector{T}: Lennard-Jones energy parameter ([ε] = J)

Mw::Vector{T}: molar mass ([Mw] = kg mol⁻¹)

collision::C: collision integral method (KimMonroe() (default) or Neufeld(), see Ω)

Constructors

ChapmanEnskogModel(components; collision_integral=KimMonroe())
ChapmanEnskogModel(components, σ, ε, Mw; collision_integral=KimMonroe(), ref="", ref_id="")

Input arguments can either be single values (pure) or vectors. In case no parameters are provided, values are taken Poling et al. (2001) or Yang et al. (2022) (if available). The keywords ref (short reference) and ref_id (DOI or ISBN) enable the specification of the reference. Mixture properties are calculated according to the models from Wilke (1950) (viscosity), Mason and Saxen (1958) (thermal conductivity), and Miller and Carman (1961) (self-diffusion coefficient).

Example

using EntropyScaling 

# Construction with custom parameters
σ, ε, Mw = 3.758e-10, 148.6*EntropyScaling.kB, 16.043e-3            # from Poling et al.
model_methane = ChapmanEnskogModel("methane",σ,ε,Mw)

η_mix = viscosity(model_methane, NaN, 300.)
D_mix = self_diffusion_coefficient(model_methane, NaN, 300.)

# Construction from database
model_mix = ChapmanEnskogModel(["butane","methanol"]; ref="Poling et al. (2001)")

η_mix = viscosity(model_mix, NaN, 300., [.5,.5])
D_mix = self_diffusion_coefficient(model_mix, NaN, 300., [.5,.5])  

References

B. E. Poling, J. M. Prausnitz, and J. P. O’Connell: The Properties of Gases and Liquids, 5th, ed. McGraw-Hill, New York (2001).

X. Yang, X. Xiao, M. Thol, M. Richter, and I. H. Bell: Linking Viscosity to Equations of State Using Residual Entropy Scaling Theory, Int. J. Thermophys. 43 (2022) 183, DOI: https://doi.org/10.1007/s10765-022-03096-9.

C. R. Wilke: A Viscosity Equation for Gas Mixtures, The Journal of Chemical Physics 18 (1950) 517–519, DOI: https://doi.org/10.1063/1.1747673.

E. A. Mason and S. C. Saxena: Approximate Formula for the Thermal Conductivity of Gas Mixtures, The Physics of Fluids 1 (1958) 361–369, DOI: https://doi.org/10.1063/1.1724352.

L. Miller and P. C. Carman: Self-Diffusion in Mixtures. Part 4. – Comparison of Theory and Experiment for Certain Gas Mixtures, Trans. Faraday Soc. 57 (1961) 2143–2150, DOI: https://doi.org/10.1039/TF9615702143.

EntropyScaling.Ω — Function

Ω(poperty::AbstractTransportProperty, model::AbstractChapmanEnskogModel, T)

Calculates the collision integral for a given model and property (Ω₁₁ for diffusion coefficients and Ω₂₂ for viscosity/thermal conductivity) at the specified temperature T.

Two methods are implemented:

KimMonroe(): Kim and Monroe (2014) and

Neufeld(): Neufeld et al. (1972)

References

S. U. Kim and C. W. Monroe: High-Accuracy Calculations of Sixteen Collision Integrals for Lennard-Jones (12-6) Gases and Their Interpolation to Parameterize Neon, Argon, and Krypton, Journal of Computational Physics 273 (2014) 358–373, DOI: https://doi.org/10.1016/j.jcp.2014.05.018.

P. D. Neufeld, A. R. Janzen, and R. A. Aziz: Empirical Equations to Calculate 16 of the Transport Collision Integrals Ω(l,s)* for the Lennard‐Jones (12–6) Potential, The Journal of Chemical Physics 57 (1972) 1100–1102, DOI: https://doi.org/10.1063/1.1678363.

Entropy Scaling Models

Entropy scaling makes use of the fact that transport properies can be scaled such that the scaled transport property $Y^{\rm s}$ is a univariate function of the configurational (or residual) entropy $s_{\rm conf}$, i.e.

\[Y^{\rm s} = Y^{\rm s}\left(s_{\rm conf}\right).\]

Entropy scaling enables the prediction of transport porperties in all fluid phases.

Models

All models are similarly structured with the following fields:

components::Vector{String}: names of the chemical components of the system

params::Vector{ModelParams}: vector of model-specific paramater objects

eos: EOS model

The ModelParams are model-specific types containing all required parameters of the model. They always contain the Chapman-Enskog model (CE_model) as well as base parameters (base) which itself contains general parameters like the transport property or the molar mass.

All models share the contructor method Model(eos, params::Dict{P}), where params is a dict containing the parameters with the respective transport property as key, e.g., Dict(Viscosity() => [a_η, b_η, c_η], ThermalConductivity() => [a_λ, b_λ, c_λ]). Here, a, b, and c are the parameters (note that a_η, b_η, ... are vectors or matrices themselfs). Lists of the parameters are given below in the repective 'Parameters' sections. Additional model-specific constructors are also given below.

EntropyScaling.FrameworkModel — Type

FrameworkModel{T} <: AbstractEntropyScalingModel

Entropy scaling framework from Schmitt et al. (2024).

The entropy scaling framework provides a physical way to model transport properties (viscosity, thermal conductivity, diffusion coeffficients) based on molecular-based EOS. It enables fitting new models using only few experimental data.

Parameters

α::Matrix{T}: component-specific parameters (size: 5 x N_components)

m (segment parameter of molecular-based EOS) and Y₀⁺min (minimum of the scaled zero-density transport property) are additional internal parameters (not to be set at construction).

Constructors

FrameworkModel(eos, params::Dict{P})

Default constructor (see above).

FrameworkModel(eos, datasets::Vector{TransportPropertyData}; 
    opts::FitOptions=FitOptions(), 
    solute=nothing
)

Constructor for fitting new parameters α to experimental data (only applicable to pure components). datasets needs to be a vector containing TransportPropertyData. opts enables controling the fitting procedure through FitOptions. solute should be an EOS model of the solute (only applicable when fitting diffusion coeffficients at infinite dilution).

Example

using EntropyScaling, Clapeyron

# Load experimental sample data for n-butane
(T_exp,ϱ_exp,η_exp) = EntropyScaling.load_sample_data()
data = ViscosityData(T_exp, [], ϱ_exp, η_exp, :unknown)

# Create EOS model
eos_model = PCSAFT("butane")

# Create entropy scaling model (fit of parameters)
model = FrameworkModel(eos_model, [data])

# Calculation of the viscostiy at state
η = viscosity(model, 0.1e6, 300.)

Reference

S. Schmitt, H. Hasse, and S. Stephan: Entropy Scaling Framework for Transport Properties Using Molecular-Based Equations of State, Journal of Molecular Liquids 395 (2024) 123811, DOI: https://doi.org/10.1016/j.molliq.2023.123811.

EntropyScaling.RefpropRESModel — Type

RefpropRESModel{T} <: AbstractEntropyScalingModel

Entropy scaling model based on Refprop EOS from Yang et al. (2022).

A database provides ready-to-use models for the viscosity of several fluids. The model can favourably be used in combination with Clapeyron.jl and Coolprop.jl (see examples).

Parameters

n::Matrix{T}: component-specific or global (group) parameters

ξ::Vector{T}: component-specific scaling parameter in case global parameters are used (ξ = 1 for individual fits)

σ::Vector{T}: LJ size parameter for the Chapman-Enskog model

ε::Vector{T}: LJ energy parameter for the Chapman-Enskog model

Constructors

RefpropRESModel(eos, params::Dict{P})

Default constructor (see above).

RefpropRESModel(eos, components)
RefpropRESModel(components)         -> only works with `Clapeyron.jl` and `Coolprop.jl`

Creates a ES model using the parameters provided in the database (recommended).

Example

using EntropyScaling, Clapeyron, CoolProp

model_pure = RefpropRESModel("R134a")
η_pure = viscosity(model_pure, 1e5, 300.; phase=:liquid)

model_mix = RefpropRESModel(["decane","butane"])
η_mix = viscosity(model_mix, 1e5, 300., [.5,.5])

References

X. Yang, X. Xiao, M. Thol, M. Richter, and I. H. Bell: Linking Viscosity to Equations of State Using Residual Entropy Scaling Theory, Int. J. Thermophys. 43 (2022) 183, DOI: https://doi.org/10.1007/s10765-022-03096-9.

Fitting Utilities

Some entropy scaling models allow the fitting of substance-specific parameters to experimental data. Therefore, a unified interface is provided including the handling of the data and the fit options.

Fitting Procedure

Loading experimental data and defining TransportPropertyData

Fitting (included in the model construction)

Plotting results and saving the parameters

EntropyScaling.TransportPropertyData — Type

TransportPropertyData(prop, T, p, ϱ, η, phase=:unknown)
ViscosityData(T, p, ϱ, η, phase=:unknown)
ThermalConductivityData(T, p, ϱ, λ, phase=:unknown)
SelfDiffusionCoefficientData(T, p, ϱ, D, phase=:unknown)
InfDiffusionCoefficientData(T, p, ϱ, D, phase=:unknown)

Constructor for TransportPropertyData. Either pressure p or density ϱ have to be specified. phase can also be a Vector{Symbol}.

Units

[T] = K

[p] = Pa

[ϱ] = mol m⁻³

[η] = Pa s

[λ] = W (m K)⁻¹

[D] = m² s⁻¹

EntropyScaling.FitOptions — Type

FitOptions

Struct to control fitting.

Fields

what_fit::Dict{AbstractTransportProperty,Vector{Bool}}: specify which parameters to fit

Example

``` FitOptions(; what_fit=Dict( ThermalConductivity()=>ones(Bool,5), SelfDiffusionCoefficient()=>Bool[0,0,0,1,1]) )