Chapman-Enskog Model

Chapman-Enskog model for transport properties at the zero-density limit based on the kinetic gas theory.

\[\begin{aligned} \eta_{\varrho \rightarrow 0} &= \frac{5}{16} \sqrt{\frac{M k_{\rm B} T}{\pi N_{\rm A}}} \frac{1}{\sigma^2 \Omega^{(2,2)}} \\ \lambda_{\varrho \rightarrow 0} &= \frac{75}{64} k_{\rm B} \sqrt{\frac{R T}{M \pi}} \frac{1}{\sigma^2 \Omega^{(2,2)}}\\ D_{\varrho \rightarrow 0} \varrho^{\rm m} &= \frac{3}{8} \sqrt{\frac{M k_{\rm B} T}{\pi N_{\rm A}}} \frac{1}{\sigma^2 \Omega^{(1,1)}} \end{aligned}\]

EntropyScaling.ChapmanEnskogModel — Type

ChapmanEnskogModel <: AbstractTransportPropertyModel

Chapman-Enskog transport properties for the zero-density limit.

Fields

σ::Vector{T}: Lennard-Jones size parameter ([σ] = m)

ε::Vector{T}: Lennard-Jones energy parameter ([ε] = J)

Mw::Vector{T}: molar mass ([Mw] = kg mol⁻¹)

collision::C: collision integral method (KimMonroe() (default) or Neufeld(), see Ω)

Constructors

ChapmanEnskogModel(components; collision_integral=KimMonroe(), ref="", ref_id=""): database constructor

ChapmanEnskogModel(components, σ, ε, Mw; collision_integral=KimMonroe()): custom parameters constructor

Input arguments can either be single values (pure) or vectors. The keywords ref (short reference) and ref_id (DOI or ISBN) enable the specification of the reference. Currently, parameters from Poling et al. [6] and Yang et al. [3] are in the database. Mixture properties are calculated according to the models from Wilke [7] (viscosity), Mason and Saxena [8] (thermal conductivity), and Miller and Carman [9] (self-diffusion).

Example

using EntropyScaling 

# Construction with custom parameters
σ, ε, Mw = 3.758e-10, 148.6*EntropyScaling.kB, 16.043e-3            # from Poling et al.
model_methane = ChapmanEnskogModel("methane",σ,ε,Mw)

η_mix = viscosity(model_methane, NaN, 300.)
D_mix = self_diffusion_coefficient(model_methane, NaN, 300.)

# Construction from database
model_mix = ChapmanEnskogModel(["butane","methanol"]; ref="Poling et al. (2001)")

η_mix = viscosity(model_mix, NaN, 300., [.5,.5])
D_mix = self_diffusion_coefficient(model_mix, NaN, 300., [.5,.5])

EntropyScaling.Ω — Function

Ω(poperty::AbstractTransportProperty, model::AbstractChapmanEnskogModel, T)

Calculates the collision integral for a given model and property (Ω₁₁ for diffusion coefficients and Ω₂₂ for viscosity/thermal conductivity) at the specified temperature T.

Two methods are implemented:

KimMonroe() [10]

Neufeld() [11]