Transport Properties

After constructing a model, transport properties can be calculated (if respective parameters are available). The call for all this is always of the form:

transport_property(model, p, T, x=[1.]; phase=:unkwown)

where

model is a AbstractTransportPropertyModel,

p is the pressure ([p] = Pa),

T is the temperature ([T] = K),

x is the mole fraction ([x] = mol mol⁻¹), and

phase is the desired phase (liquid or gas) used in the volume solver.

Internally, the density is always calculated first and then the function EntropyScaling.ϱT_transport_property(model, ϱ, T, x) is called (with [ϱ] = mol m⁻³).

EntropyScaling.viscosity — Function

viscosity(model::EntropyScalingModel, p, T, z=[1.]; phase=:unknown)

Viscosity η(p,T,x) ([η] = Pa s).

EntropyScaling.thermal_conductivity — Function

thermal_conductivity(model::EntropyScalingModel, p, T, z=[1.]; phase=:unknown)

Thermal conductivity λ(p,T,x) ([λ] = W m⁻¹ K⁻¹).

EntropyScaling.self_diffusion_coefficient — Function

self_diffusion_coefficient(model::EntropyScalingModel, p, T, z=[1.]; phase=:unknown)

Self-diffusion coefficient D(p,T,x) ([D] = m² s⁻¹).

EntropyScaling.MS_diffusion_coefficient — Function

MS_diffusion_coefficient(model::EntropyScalingModel, p, T, z; phase=:unknown)

Maxwell-Stefan diffusion coefficient Ð(p,T,x) ([Ð] = m² s⁻¹).