Models

EntropyScaling.ChapmanEnskogModel

EntropyScaling.FrameworkModel

EntropyScaling.RefpropRESModel

EntropyScaling.Ω

Chapman-Enskog Models

Chapman-Enskog model the for transport properties at the zero-density limit based on the kinetic gas theory.

\[\begin{aligned} \eta_{\varrho \rightarrow 0} &= \frac{5}{16} \sqrt{\frac{M k_{\rm B} T}{\pi N_{\rm A}}} \frac{1}{\sigma^2 \Omega^{(2,2)}} \\ \lambda_{\varrho \rightarrow 0} &= \frac{75}{64} k_{\rm B} \sqrt{\frac{R T}{M \pi}} \frac{1}{\sigma^2 \Omega^{(2,2)}}\\ D_{\varrho \rightarrow 0} \varrho^{\rm m} &= \frac{3}{8} \sqrt{\frac{M k_{\rm B} T}{\pi N_{\rm A}}} \frac{1}{\sigma^2 \Omega^{(1,1)}} \end{aligned}\]

EntropyScaling.ChapmanEnskogModel — Type

ChapmanEnskogModel <: AbstractTransportPropertyModel

Chapman-Enskog transport properties for the zero-density limit.

Fields

σ::Vector{T}: Lennard-Jones size parameter ([σ] = m)

ε::Vector{T}: Lennard-Jones energy parameter ([ε] = J)

Mw::Vector{T}: molar mass ([Mw] = kg mol⁻¹)

collision::C: collision integral method (KimMonroe() (default) or Neufeld(), see Ω)

Constructors

ChapmanEnskogModel(components; collision_integral=KimMonroe())*  
ChapmanEnskogModel(components, σ, ε, Mw; collision_integral=KimMonroe())

Input arguments can either be single values (pure) or vectors. In case no parameters are provided, values are taken Poling et al. (2001) or Yang et al. (2022) (if available). Mixture properties are calculated according to the models from Wilke (1950) (viscosity), Mason and Saxen (1958) (thermal conductivity), and Miller and Carman (1961) (self-diffusion coefficient). [* to be implemented]

References

B. E. Poling, J. M. Prausnitz, and J. P. O’Connell: The Properties of Gases and Liquids, 5th, ed. McGraw-Hill, New York (2001).

X. Yang, X. Xiao, M. Thol, M. Richter, and I. H. Bell: Linking Viscosity to Equations of State Using Residual Entropy Scaling Theory, Int. J. Thermophys. 43 (2022) 183, DOI: https://doi.org/10.1007/s10765-022-03096-9.

C. R. Wilke: A Viscosity Equation for Gas Mixtures, The Journal of Chemical Physics 18 (1950) 517–519, DOI: https://doi.org/10.1063/1.1747673.

E. A. Mason and S. C. Saxena: Approximate Formula for the Thermal Conductivity of Gas Mixtures, The Physics of Fluids 1 (1958) 361–369, DOI: https://doi.org/10.1063/1.1724352.

L. Miller and P. C. Carman: Self-Diffusion in Mixtures. Part 4. – Comparison of Theory and Experiment for Certain Gas Mixtures, Trans. Faraday Soc. 57 (1961) 2143–2150, DOI: https://doi.org/10.1039/TF9615702143.

EntropyScaling.Ω — Function

Ω(poperty::AbstractTransportProperty, model::AbstractChapmanEnskogModel, T)

Calculates the collision integral for a given model and property (Ω₁₁ for diffusion coefficients and Ω₂₂ for viscosity/thermal conductivity) at the specified temperature T.

Two methods are implemented:

KimMonroe(): Kim and Monroe (2014) and

Neufeld(): Neufeld et al. (1972)

References

S. U. Kim and C. W. Monroe: High-Accuracy Calculations of Sixteen Collision Integrals for Lennard-Jones (12-6) Gases and Their Interpolation to Parameterize Neon, Argon, and Krypton, Journal of Computational Physics 273 (2014) 358–373, DOI: https://doi.org/10.1016/j.jcp.2014.05.018.

P. D. Neufeld, A. R. Janzen, and R. A. Aziz: Empirical Equations to Calculate 16 of the Transport Collision Integrals Ω(l,s)* for the Lennard‐Jones (12–6) Potential, The Journal of Chemical Physics 57 (1972) 1100–1102, DOI: https://doi.org/10.1063/1.1678363.

Entropy Scaling Models

Entropy scaling makes use of the fact that transport properies can be scaled such that the scaled transport property $Y^{\rm s}$ is a univariate function of the configurational (or residual) entropy $s_{\rm conf}$, i.e.

\[Y^{\rm s} = Y^{\rm s}\left(s_{\rm conf}\right).\]

Entropy scaling enables the prediction of transport porperties in all fluid phases.

Available Models

EntropyScaling.FrameworkModel — Type

FrameworkModel{T} <: AbstractEntropyScalingModel

A generic entropy scaling model.

EntropyScaling.RefpropRESModel — Type

RefpropRESModel{T} <: AbstractEntropyScalingModel

Entropy scaling model based on Refprop EOS.

References

Fitting Entropy Scaling Parameters

[Work in progress]

Some entropy scaling models allow the fitting of substance-specific parameters to experimental data. Therefore, a unified interface is provided including the handling of the data and the fit options.

Fitting Procedure

Loading experimental data

Fitting

Plotting results and saving the parameters

Example